% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactsnopar.R
\name{impactsnopar}
\alias{impactsnopar}
\title{Compute direct, indirect and total impacts functions 
  for continous non-parametric covariates in semiparametric spatial regression models.}
\usage{
impactsnopar(
  obj,
  listw = NULL,
  conflevel = 0.95,
  viewplot = TRUE,
  smooth = TRUE,
  span = c(0.1, 0.1, 0.2)
)
}
\arguments{
\item{obj}{\emph{pspatfit} object fitted using \code{\link{pspatfit}} function.}

\item{listw}{should be a spatial neighbours list object created for example by \code{nb2listw} from \code{spdep} package. 
It can also be a spatial weighting matrix of order (NxN) instead of a listw neighbours list object.}

\item{conflevel}{numerical value for the confidence interval of the
impact functions. Default 0.95.}

\item{viewplot}{Default \code{TRUE} to plot impacts. If FALSE use \code{\link{plot_impactsnopar}} to plot impacts}

\item{smooth}{Default \code{TRUE}. Whether to smooth fitted impacts or not.}
}
\value{
A list including
  \tabular{ll}{
    \emph{impnopar_tot} \tab Matrix including total impects in columns. \cr
    \emph{impnopar_dir} \tab Matrix including direct impects in columns. \cr
    \emph{impnopar_ind} \tab Matrix including indirect impects in columns. \cr
    \emph{impnopar_tot_up} \tab Matrix including upper bounds of total impects in columns. \cr
    \emph{impnopar_dir_up} \tab Matrix including upper bounds of direct impects in columns. \cr
    \emph{impnopar_ind_up} \tab Matrix including upper bounds of indirect impects in columns. \cr
    \emph{impnopar_tot_low} \tab Matrix including lower bounds of total impects in columns. \cr
    \emph{impnopar_dir_low} \tab Matrix including lower bounds of direct impects in columns. \cr
    \emph{impnopar_ind_low} \tab Matrix including lower bounds of indirect impects in columns. \cr
 }
}
\description{
Compute and plot direct, indirect and total impect functions for 
  non-parametric covariates included in a semiparametric spatial
  or spatio-temporal SAR model. This model must include a spatial
  lag of the dependent variable (SAR) to have indirect impacts 
  different from 0, otherwise, total and direct function impacts 
  are the same.
}
\details{
DESCRIBE ALGORITHM TO COMPUTE impECT FUNCTIONS AND THE 
         SMOOTHING TO PLOT
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(pspatreg)
data(unemp_it); Wsp <- Wsp_it

######################  No Spatial Trend: PSAR including a spatial 
######################  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20) 
 gamsar <- pspatfit(form1, data = unemp_it, type = "sar", listw = Wsp_it)
 summary(gamsar)
 ###### Non-Parametric Total, Direct and Indirect impects
 eff_nparvar <- impactsnopar(gamsar, listw = Wsp_it, viewplot = TRUE)
 
######################   PSAR-ANOVA with spatial trend
form2 <- unrate ~ partrate + agri + cons +
                  pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                  pspt(long,lat,nknots=c(20,20),psanova=TRUE,
                  nest_sp1=c(1,2),nest_sp2=c(1,2))
##### Spatial trend fixed for period 1996-2014
geospanova_sar <- pspatfit(form2, data=unemp_it, listw = Wsp_it, type = "sar", 
                           control=list(tol=1e-1, maxit=200, trace=FALSE))
summary(geospanova_sar)
 ###### Non-Parametric Total, Direct and Indirect impects
 eff_nparvar2 <- impactsnopar(geospanova_sar, listw = Wsp_it, viewplot = TRUE)
 
######################   PSAR-ANOVA with spatio-temporal trend and 
######################   temporal autorregresive noise
 form3 <- unrate ~ partrate + agri + cons +
                   pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                   pspt(long,lat,year,nknots=c(18,18,8),psanova=TRUE,
                   nest_sp1=c(1,2,3),nest_sp2=c(1,2,3),
                   nest_time=c(1,2,2),ntime=19)
sptanova_sar_ar1 <- pspatfit(form3, data = unemp_it, Wsp = Wsp_it, type = "sar", ar1=TRUE,
                    control=list(thr=1e-1, maxit=200, trace=FALSE))
summary(sptanova_sar_ar1)
 ###### Non-Parametric Total, Direct and Indirect impects
 eff_nparvar3 <- impactsnopar(geospanova_sar, listw = Wsp_it, viewplot = TRUE)

}
\references{
\itemize{ 
  \item Basile, R., Durbán, M., Mínguez, R., Montero, J.
        M., and Mur, J. (2014). Modeling regional economic 
        dynamics: Spatial dependence, spatial heterogeneity and 
        nonlinearities. \emph{Journal of Economic Dynamics and 
        Control}, (48), 229-245.
        
  \item LeSage, J. and Pace, K. (2009). \emph{Introduction to 
        Spatial Econometrics}. CRC Press, Boca Raton.
        
  \item Mínguez, R.; Basile, R. and Durbán, M. (2018). An Alternative Semiparametric Model
        for Spatial Panel Data. Under evaluation in \emph{Statistical
        Methods and Applications}.
 }
}
\seealso{
\itemize{
  \item \code{\link{pspatfit}} estimate spatial or spatio-temporal semiparametric PS-SAR
    regression models.
  \item \code{\link{impactspar}} compute and simulate total, direct and indirect impect
    (or impacts) for parametric continuous covariates.
  \item \code{\link{fit_terms}} compute terms for smooth functions for non-parametric
    continuous covariates and for non-parametric trends.
  \item \code{\link{plot_impactsnopar}} plot the non-parametric impects functions
    allowing for previous smoothing.
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
\concept{Direct, Indirect and Total impects.}
\keyword{Direct}
\keyword{Indirect}
\keyword{SAR,}
\keyword{covariates.}
\keyword{impects,}
\keyword{non-parametric}
