% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactspar.R
\name{impactspar}
\alias{impactspar}
\title{Compute direct, indirect and total impacts
  for continous parametric covariates.}
\usage{
impactspar(obj, ..., tr = NULL, R = 1000, listw = NULL, tol = 1e-06, Q = NULL)
}
\arguments{
\item{obj}{A `pspatreg` object created by \code{\link{pspatfit}}.}

\item{...}{Arguments passed through to methods in the \pkg{coda} package}

\item{tr}{A vector of traces of powers of the spatial weights matrix created using \code{trW}, for approximate impact measures; if not given, \code{listw} must be given for exact measures (for small to moderate spatial weights matrices); the traces must be for the same spatial weights as were used in fitting the spatial regression, and must be row-standardised}

\item{R}{If given, simulations are used to compute distributions for the impact measures, returned as \code{mcmc} objects; the objects are used for convenience but are not output by an MCMC process}

\item{listw}{If \code{tr} is not given, a spatial weights object as created by \code{nb2listw}; they must be the same spatial weights as were used in fitting the spatial regression, but do not have to be row-standardised}

\item{tol}{Argument passed to \code{mvrnorm}: tolerance (relative to largest variance) for numerical lack of positive-definiteness in the coefficient covariance matrix}

\item{Q}{default NULL, else an integer number of cumulative power series impacts to calculate if \code{tr} is given}
}
\value{
An object of class \emph{eff.par.pspat}. Can be printed
        with \code{summary}.
        
        The object returned is a list with 3 matrices including
        the results of simulated effects:
         \tabular{ll}{
           \emph{tot_eff} \tab Matrix including simulated total effects 
                               for each variable in rows. \cr
           \emph{dir_eff} \tab Matrix including simulated direct effects 
                               for each variable in rows. \cr
           \emph{ind_eff} \tab Matrix including simulated indirect effects 
                               for each variable in rows. \cr
         }
}
\description{
Compute direct, indirect and total 
  impacts for parametric covariates included in a semiparametric spatial
  or spatio-temporal model.
}
\details{
DESCRIBE ALGORITHM TO SIMULATE PARAMETRIC EFFECTS
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(pspatreg)
data(unemp_it, package = "pspatreg")
## Wsp_it is a matrix. Create a neighboord list 
lwsp_it <- spdep::mat2listw(Wsp_it)
## short sample for spatial pure case (2d)
unemp_it_short <- unemp_it[unemp_it$year == 2019, ]  
########  No Spatial Trend: PSAR including a spatial 
########  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv, nknots = 15) +
                 pspl(empgrowth, nknots = 20) 
gamsar <- pspatfit(form1, 
                   data = unemp_it_short, 
                   type = "sar", 
                   listw = lwsp_it)
summary(gamsar)
###### Parametric Total, Direct and Indirect Effects
imp_parvar <- impactspar(gamsar, listw = lwsp_it)
summary(imp_parvar)
\donttest{
######################   PSAR-ANOVA with spatio-temporal trend and 
######################   temporal autorregresive noise
form2 <- unrate ~ partrate + agri + cons +
                  pspl(serv, nknots = 15) + 
                  pspl(empgrowth, nknots = 20) +
                  pspt(long, lat, year, 
                       nknots = c(18, 18, 8), 
                       psanova = TRUE,
                       nest_sp1 = c(1, 2, 3),
                       nest_sp2 = c(1, 2, 3),
                       nest_time = c(1, 2, 2))
sptanova_sar_ar1 <- pspatfit(form2, 
                             data = unemp_it, 
                             listw = lwsp_it, 
                             type = "sar", 
                             cor = "ar1", 
                             control = list(tol = 1e-1))
summary(sptanova_sar_ar1)
##### Parametric Total, Direct and Indirect Effects
imp_parvar2 <- impactspar(sptanova_sar_ar1, 
                          listw = lwsp_it)
summary(imp_parvar2)
}

}
\references{
\itemize{ 
  \item LeSage, J. and Pace, K. (2009). \emph{Introduction to 
        Spatial Econometrics}. CRC Press, Boca Raton.
 }
}
\seealso{
\itemize{
  \item \code{\link{pspatfit}} estimate spatial or spatio-temporal 
          semiparametric PS-SAR, PS-SEM, PS-SARAR, PS-SLX or PS-DURBIN regression models.
  \item \code{\link{impactsnopar}} compute total, direct and indirect effect
          functions for non-parametric continuous covariates.
  \item \code{\link{fit_terms}} compute smooth functions for non-parametric
          continuous covariates.
  \item \code{\link[spdep]{impacts}} similar function in \pkg{spdep} 
          package to compute impacts in spatial parametric econometric 
          models.                              
}
}
\concept{Direct, Indirect and Total Effects.}
\keyword{Direct}
\keyword{Indirect}
\keyword{covariates.}
\keyword{effects,}
\keyword{parametric}
