% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{map_it}
\alias{map_it}
\title{map of Italian provinces}
\format{
An sf object with 103 rows and 2 columns:
\describe{
  \item{COD_PRO}{province (NUTS-3) coded as a number.}
  \item{geometry}{geometry (polygons) of the sf object.}
 }
}
\source{
Italian National Institute of Statistics (ISTAT)
        \emph{https://www.istat.it}
}
\usage{
map_it
}
\description{
An sf object including a map of Italian NUTS-3 provinces
}
\keyword{datasets}
