% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sp3d.R
\name{plot_sp3d}
\alias{plot_sp3d}
\title{Plot and mapping spatio-temporal trends}
\usage{
plot_sp3d(object, data, time_var, time_index, addmain = TRUE, addint = TRUE)
}
\arguments{
\item{object}{object returned from \code{\link{pspatfit}}}

\item{data}{sf object.}

\item{time_var}{name of the temporal variable in data.}

\item{time_index}{vector of time points to plot.}

\item{addmain}{Add f1_main and f2_main plots in psanova case.}

\item{addint}{Add f12_int in psanova case.}

\item{coordinates}{coordinates matrix if *data* is not an sf object.}

\item{npoints}{number of points to use in the interpolation.}

\item{addcontour}{Logical value to add contour lines.}

\item{addpoints}{Logical value to add spatial points to the graphics.}
}
\value{
plots and maps of the spatial trends
}
\description{
DESCRIBE THE FUNCTION...
}
\examples{

library(pspatreg)
data("unemp_it_new_polyg_sf")
data("unemp_it")
unemp_it <- unemp_it_new_polyg_sf
rm(unemp_it_new_polyg_sf) # Only to be consistent with the names...
colnames(Wsp_it) <- rownames(Wsp_it)
Wsp_it <- as.matrix(Wsp_it)
Using spdep we transform the spatial weights matrix in a list of neighbours object:
 
 lwsp_it <- spdep::mat2listw(Wsp_it, 
                             row.names = rownames(Wsp_it))

summary(lwsp_it)
######## FORMULA of the model
form3d_psanova_restr <- unrate ~ partrate + agri + cons +
pspl(serv, nknots = 15) + 
 pspl(empgrowth, nknots = 20) +
 pspt(long, lat, year, 
      nknots = c(18,18,8),
      psanova = TRUE, 
      nest_sp1 = c(1, 2, 3), 
      nest_sp2 = c(1, 2, 3),
      nest_time = c(1, 2, 2),
      f1t = FALSE, f2t = FALSE)

####### FIT the model
sp3danovasarar1 <- pspatfit(form3d_psanova_restr, 
data = unemp_it, 
listw = lwsp_it, 
method = "Chebyshev", 
type = "sar",
cor = "ar1")

###### PLOT spatio-temporal trends
plot_sp3d(sp3danovasarar1, data = unemp_it, 
time_var = "year", 
time_index = c(1996, 2005, 2019),
addmain = FALSE, addint = FALSE)

}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
