% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sptime.R
\name{plot_sptime}
\alias{plot_sptime}
\title{Plot of time trend for PS-ANOVA models in 3d.}
\usage{
plot_sptime(object, data, time_var, reg_var)
}
\arguments{
\item{object}{object returned from \code{\link{pspatfit}}}

\item{data}{either sf or dataframe with the data.}

\item{time_var}{name of the temporal variable in data.}

\item{reg_var}{name of the regional variable in data.}
}
\value{
time series plots of the temporal trend for each region
}
\description{
DESCRIBE THE FUNCTION...
}
\examples{

library(pspatreg)
data("unemp_it_new_polyg_sf")
data("unemp_it")
unemp_it <- unemp_it_new_polyg_sf
rm(unemp_it_new_polyg_sf) # Only to be consistent with the names...
colnames(Wsp_it) <- rownames(Wsp_it)
Wsp_it <- as.matrix(Wsp_it)
Using spdep we transform the spatial weights matrix in a list of neighbours object:
 
 lwsp_it <- spdep::mat2listw(Wsp_it, 
                             row.names = rownames(Wsp_it))

summary(lwsp_it)

###### FORMULA OF THE MODEL
form3d_psanova_restr <- unrate ~ partrate + agri + cons +
pspl(serv, nknots = 15) + 
 pspl(empgrowth, nknots = 20) +
 pspt(long, lat, year, 
      nknots = c(18,18,8),
      psanova = TRUE, 
      nest_sp1 = c(1, 2, 3), 
      nest_sp2 = c(1, 2, 3),
      nest_time = c(1, 2, 2),
      f1t = FALSE, f2t = FALSE)

####### FIT the model
sp3danovasarar1 <- pspatfit(form3d_psanova_restr, 
data = unemp_it, 
listw = lwsp_it, 
method = "Chebyshev", 
type = "sar",
cor = "ar1") 

######## PLOT OF TEMPORAL TREND FOR EACH PROVINCE 
plot_sptime(sp3danovasarar1, data = unemp_it, 
time_var = "year", 
reg_var = "prov")

}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
