% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.pspatreg.R
\name{print.summary.pspatreg}
\alias{print.summary.pspatreg}
\title{Print method for objects of class summary.pspatreg.}
\usage{
\method{print}{summary.pspatreg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \emph{summary.pspatreg}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for objects of class summary.pspatreg.
}
\examples{
# See examples for \code{\link{pspatfit}} function.
}
\seealso{
\itemize{
  \item \code{\link{summary.pspatreg}} Summary method for \emph{pspatreg} objects.
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
