% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{prod_it}
\alias{prod_it}
\title{Productivity growth and internal net migration - Italian provinces}
\format{
An sf object with 107 rows and 9 columns:
\describe{
  \item{COD_PROV}{province (NUTS-3) coded as a number.}
  \item{DEN_PROV}{province (NUTS-3) coded as a name.}
  \item{longitude}{longitude of the centroid of the province.}
  \item{latitude}{latitude of the centroid of the province.}
  \item{lnPROD_0}{log of labor productivity in 2002
                  (measured as gross value added per worker).}
  \item{growth_PROD}{Average annual growth rate of labor productivity 
                     over the period 2002-2018.}
  \item{lnoccgr}{Average annual growth rate of employment over the period 2002-2018.}
  \item{net}{Average annual provincial internal net migration rate 
             (computed as the difference between internal immigration and 
             emigration flows of the working-age population, i.e. 
             people aged 15-65, divided by the total population).}
  \item{geometry}{geometry (polygons) of the sf object.}
}
}
\source{
Italian National Institute of Statistics (ISTAT)
        \emph{https://www.istat.it}
}
\usage{
prod_it
}
\description{
A spatial dataframe including a map of Italian NUTS-3 provinces and 
cross-sectional dataset on provincial labor productivity growth rates,
internal net migration rates, and other economic variables.
}
\keyword{datasets}
