% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspl_functions.R
\name{pspl}
\alias{pspl}
\title{Non-parametric terms for continuous covariates in 
  semiparametric PS-SAR, PS-SEM, PS-SARAR, PS-SLX or PS-DURBIN models.}
\usage{
pspl(
  x,
  xl = min(x) - 0.01,
  xr = max(x) + 0.01,
  nknots = 10,
  bdeg = 3,
  pord = 2,
  decom = 2
)
}
\arguments{
\item{x}{Name of the covariate.}

\item{xl}{Minimum of the interval for the continuous covariate.
Default = min(x) - 0.01.}

\item{xr}{Maximum of the interval for the continuous covariate.
Default = max(x) - 0.01.}

\item{nknots}{Number of knots for spline basis. Default = 10.}

\item{bdeg}{Order of the B-spline basis. Default = 3.}

\item{pord}{Order of the penalty for the difference matrix 
in P-spline. Default = 2.}

\item{decom}{Type of decomposition of fixed part when P-spline
term is expressed as a mixed model. If \code{decom = 1} the 
fixed part is given by \eqn{X = B*U_n} where \emph{B} is the
B-spline basis matrix and \emph{U_n} is the nullspace basis of the 
penalty matrix. If \code{decom = 2} the fixed part is given by
\eqn{X = [1|x|...|x^(pord-1)] }. Default = 2.

@return A list object.
 \tabular{ll}{
   \code{B} \tab Matrix including B-spline basis for the covariate \cr
   \code{a} \tab List including \emph{nknots}, \emph{knots}, \emph{bdeg},
      \emph{pord} and \emph{decom}.  \cr
 }}
}
\description{
This function allows the inclusion of non-parametric
  continuous covariates in the formula of semiparametric PS-SAR, 
  PS-SEM, PS-SARAR, PS-SLX or PS-DURBIN models. Each non-parametric 
  covariate must be included with its own \code{pspl} term in a
 formula. See \code{Examples} to understand how to do this. Each 
 non-parametric term is modelled using p-splines.
}
\examples{
For more in-depth examples see \code{\link{pspatfit}}

library(pspatreg)
## load spatial panel and Wsp_it
## 103 Italian provinces. Period 1996-2019
data(unemp_it, package = "pspatreg")
## Wsp_it is a matrix. Create a neighboord list 
lwsp_it <- spdep::mat2listw(Wsp_it)
## short sample for spatial pure case (2d)
unemp_it_short <- unemp_it[unemp_it$year == 2019, ]
####  GAM pure with pspatreg
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv, nknots = 15) +
                 pspl(empgrowth, nknots = 20)
gampure <- pspatfit(form1, data = unemp_it_short)
summary(gampure)

}
\seealso{
\itemize{
  \item \code{\link{pspt}} Include a non-parametric spatial or 
    spatio-temporal trend in the formula of the model.
}
}
