% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pspatreg.R
\name{summary.pspatreg}
\alias{summary.pspatreg}
\title{Summary method for objects of class pspatreg.}
\usage{
\method{summary}{pspatreg}(object, ...)
}
\arguments{
\item{object}{\emph{pspatreg} object fitted using \code{\link{pspatreg}} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \emph{summary.pspatreg}
}
\description{
This method summarizes both spatial (2-dimension) and
  spatio-temporal (3-dimension) \emph{pspatreg} objects.
  The tables include information of:
  \itemize{
     \item The spatial (or spatio-temporal) trends. When the model is ANOVA
       the trend is decomposed in main and interaction effects.
     \item The parametric and non-parametric covariates.
     \item The \eqn{\rho} parameter when the model is SAR.
     \item The \eqn{\phi} parameter when the model is spatio-temporal
       with a first-order autorregressive in the noise.
 }
}
\examples{
 # See examples for \code{\link{pspatfit}} function.

}
\seealso{
\itemize{
  \item \code{\link{pspatfit}} estimate spatial or spatio-temporal semiparametric PS-SAR, PS-SEM, PS-SARAR, PS-SLX or PS-DURBIN
  regression models.
  \item \code{\link{print.summary.pspatreg}} print objects of class \emph{summary.pspatreg}
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
