% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Wsp_it}
\alias{Wsp_it}
\title{Spatial weight matrix for Italian provinces}
\format{
A row-standardized squared matrix with 103 rows and columns. 
The rows and columns follow the same order than provinces included in 
\emph{unemp_it} data frame.
}
\source{
Italian National Institute of Statistics (ISTAT)
        \url{https://www.istat.it/}
}
\usage{
Wsp_it
}
\description{
A spatial weight matrix row-standardized for Italian NUTS-3 provinces
}
\keyword{datasets}
