% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_pspatreg.R
\name{methods_pspatreg}
\alias{methods_pspatreg}
\alias{anova}
\alias{anova.pspatreg}
\alias{coef}
\alias{coef.pspatreg}
\alias{fitted}
\alias{fitted.pspatreg}
\alias{logLik}
\alias{logLik.pspatreg}
\alias{residuals}
\alias{residuals.pspatreg}
\alias{vcov}
\alias{vcov.pspatreg}
\alias{print}
\alias{print.pspatreg}
\title{Methods for class pspatreg}
\usage{
\method{anova}{pspatreg}(object, ..., lrtest = TRUE)

\method{coef}{pspatreg}(object, ...)

\method{fitted}{pspatreg}(object, ...)

\method{logLik}{pspatreg}(object, ..., REML = FALSE)

\method{residuals}{pspatreg}(object, ...)

\method{vcov}{pspatreg}(object, ..., bayesian = FALSE)

\method{print}{pspatreg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{a `pspatreg` object created by 
\code{\link{pspatfit}}.}

\item{...}{further arguments passed to or from other methods.}

\item{lrtest}{logical value to compute likelihood ratio
test for nested models in `anova` method. Default = `TRUE`}

\item{REML}{logical value to get restricted log-likelihood 
instead of the usual log-likelihood. Default = `FALSE`}

\item{bayesian}{logical value to get bayesian or frequentist
covariance matrix for parametric terms. Default = `FALSE`}

\item{x}{similar to \code{object} argument for \code{print()} 
and \code{plot} functions.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}
}
\description{
The \code{\link{anova}} function provides tables of fitted 
  \code{pspatreg} models including information criteria (AIC and BIC), 
  log-likelihood and degrees of freedom of each fitted model. The 
  argument `lrtest` allows to perform LR tests between nested models.
  The \code{\link{print}} function is used to print short tables including the 
  values of beta and spatial coefficients as well as p-values of significance test for each 
  coefficient. This can be used as an alternative to 
 \code{\link{summary.pspatreg}} when a brief output is needed. 
 The rest of methods works in the usual way.
}
\examples{
## INCLUDE SOME EXAMPLES OF ANOVA AND PERHAPS ANOTHER 
## ADDITIONAL METHODS
}
\author{
\tabular{ll}{
    Roman Minguez \tab \email{roman.minguez@uclm.es} \cr
    Roberto Basile \tab \email{roberto.basile@univaq.it} \cr
    Maria Durban \tab \email{mdurban@est-econ.uc3m.es} \cr
    Gonzalo Espana-Heredia \tab \email{gehllanza@gmail.com} \cr
  }
}
