% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impactsnopar.R
\name{plot_impactsnopar}
\alias{plot_impactsnopar}
\title{Plot direct, indirect and total impacts  functions 
  for continous non-parametric covariates in semiparametric spatial regression models.}
\usage{
plot_impactsnopar(
  impactsnopar,
  data,
  smooth = TRUE,
  span = c(0.1, 0.1, 0.2),
  dynamic = FALSE,
  nt = NULL
)
}
\arguments{
\item{impactsnopar}{object returned from \code{\link{impactsnopar}} function.}

\item{data}{dataframe with the data.}

\item{smooth}{logical value to choose smoothing of the effects function
prior to plot. Default TRUE.}

\item{span}{span for the kernel of the smoothing (see \code{\link{loess}} 
for details). Default c(0.1,0.1,0.2).}
}
\value{
plot of the direct, indirect and total impacts  function for each non-parametric
  covariate included in the object returned from \code{\link{impactsnopar}}.
}
\description{
Plot direct, indirect and total impacts functions for 
  non-parametric covariates included in a semiparametric spatial
  or spatio-temporal SAR model. This model must include a spatial
  lag of the dependent variable (SAR) to have indirect effects 
  different from 0, otherwise, total and direct function effects 
  are the same. The effect functions can be smoothed to overcome 
  the instabilities created by the premultiplication of matrix
  \eqn{(I - \rho W)^{-1}}
}
\examples{
################################################
 ###################### Examples using a panel data of rate of
 ###################### unemployment for 103 Italian provinces in period 1996-2014.
library(pspatreg)
data(unemp_it); Wsp <- Wsp_it

######################  No Spatial Trend: PSAR including a spatial 
######################  lag of the dependent variable
form1 <- unrate ~ partrate + agri + cons +
                 pspl(serv,nknots=15) +
                 pspl(empgrowth,nknots=20) 
 gamsar <- pspatfit(form1, data = unemp_it, type = "sar", listw = Wsp_it)
 summary(gamsar)
 ###### Non-Parametric Total, Direct and Indirect impects
 eff_nparvar <- impactsnopar(gamsar, listw = Wsp_it, viewplot = TRUE)
 
 ##### This returns the same result but using plot_impactsnopar()
 eff_nparvar <- impactsnopar(gamsar, listw = Wsp_it, viewplot = FALSE)
 plot_impactsnopar(eff_nparvar, data = unemp_it, smooth = TRUE)
 
######################   PSAR-ANOVA with spatial trend
form2 <- unrate ~ partrate + agri + cons +
                  pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                  pspt(long,lat,nknots=c(20,20),psanova=TRUE,
                  nest_sp1=c(1,2),nest_sp2=c(1,2))
##### Spatial trend fixed for period 1996-2014
geospanova_sar <- pspatfit(form2, data=unemp_it, listw = Wsp_it, type = "sar", 
                           control=list(thr=1e-1, maxit=200, trace=FALSE))
summary(geospanova_sar)
 ###### Non-Parametric Total, Direct and Indirect impects
 eff_nparvar2 <- impactsnopar(geospanova_sar, listw = Wsp_it, viewplot = FALSE)
 plot_impactsnopar(eff_nparvar2, data = unemp_it, smooth = TRUE)
 
######################   PSAR-ANOVA with spatio-temporal trend and 
######################   temporal autorregresive noise
 form3 <- unrate ~ partrate + agri + cons +
                   pspl(serv,nknots=15) + pspl(empgrowth,nknots=20) +
                   pspt(long,lat,year,nknots=c(18,18,8),psanova=TRUE,
                   nest_sp1=c(1,2,3),nest_sp2=c(1,2,3),
                   nest_time=c(1,2,2),ntime=19)
sptanova_sar_ar1 <- pspatfit(form3, data = unemp_it, Wsp = Wsp_it, type = "sar", ar1=TRUE,
                    control=list(thr=1e-1, maxit=200, trace=FALSE))
summary(sptanova_sar_ar1)
 ###### Non-Parametric Total, Direct and Indirect impects
 eff_nparvar3 <- impactsnopar(geospanova_sar, listw = Wsp_it, viewplot = TRUE)
 plot_impactsnopar(eff_nparvar3, data = unemp_it, smooth = TRUE)

}
\references{
\itemize{ 
  \item Basile, R., Durbán, M., Mínguez, R., Montero, J.
        M., and Mur, J. (2014). Modeling regional economic 
        dynamics: Spatial dependence, spatial heterogeneity and 
        nonlinearities. \emph{Journal of Economic Dynamics and 
        Control}, (48), 229-245.                  
 }
}
\seealso{
\itemize{
  \item \code{\link{impactsnopar}} compute total, direct and indirect effect
          functions for non-parametric continuous covariates.
  \item \code{\link{fit_terms}} compute smooth functions for non-parametric
                                continuous covariates.
  \item \code{\link{plot_terms}} plot the terms of non-parametric covariates.
}
}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
\concept{Direct, Indirect and Total Impacts.}
\keyword{Direct}
\keyword{Indirect}
\keyword{SAR,}
\keyword{covariates.}
\keyword{effects,}
\keyword{impacts,}
\keyword{non-parametric}
