% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sp3d.R
\name{plot_sp3d}
\alias{plot_sp3d}
\title{Plot and mapping spatio-temporal trends}
\usage{
plot_sp3d(object, data, time_var, time_index, addmain = TRUE, addint = TRUE)
}
\arguments{
\item{object}{object returned from \code{\link{pspatfit}}}

\item{data}{sf object.}

\item{time_var}{name of the temporal variable in data.}

\item{time_index}{vector of time points to plot.}

\item{addmain}{Add f1_main and f2_main plots in psanova case.}

\item{addint}{Add f12_int in psanova case.}
}
\value{
plots and maps of the spatial trends
}
\description{
DESCRIBE THE FUNCTION...
}
\examples{

library(pspatreg)
library(sf)
data("unemp_it")
lwsp_it <- spdep::mat2listw(Wsp_it, row.names = rownames(Wsp_it))
unemp_it_sf <- st_as_sf(dplyr::left_join(
                                unemp_it, map_it,  
                        by = c("prov" = "COD_PRO")))
######## FORMULA of the model
form3d_psanova_restr <- unrate ~ partrate + agri + cons +
pspl(serv, nknots = 15) + 
 pspl(empgrowth, nknots = 20) +
 pspt(long, lat, year, 
      nknots = c(18,18,8),
      psanova = TRUE, 
      nest_sp1 = c(1, 2, 3), 
      nest_sp2 = c(1, 2, 3),
      nest_time = c(1, 2, 2),
      f1t = FALSE, f2t = FALSE)

####### FIT the model
sp3danova <- pspatfit(form3d_psanova_restr, 
                         data = unemp_it_sf)
summary(sp3danova)                          

###### Plot spatio-temporal trends for different years
plot_sp3d(sp3danova, data = unemp_it_sf, 
          time_var = "year", 
          time_index = c(1996, 2005, 2019),
          addmain = FALSE, addint = FALSE)
###### Plot of spatio-temporal trend, main effects 
######      and interaction effect for a year
plot_sp3d(sp3danova, data = unemp_it_sf, 
          time_var = "year", 
          time_index = c(2019),
          addmain = TRUE, addint = TRUE)

}
\author{
Roman Minguez \email{roman.minguez@uclm.es}
}
