% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.impactspar.pspatreg.R
\name{print.summary.impactspar.pspatreg}
\alias{print.summary.impactspar.pspatreg}
\title{Print method for objects of class summary.impactspar.pspatreg}
\usage{
\method{print}{summary.impactspar.pspatreg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \emph{summary.impactspar.pspatreg}.}

\item{digits}{number of digits to show in printed tables.
Default: max(3L, getOption("digits") - 3L).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
No return value, called for side effects.
}
\description{
Print method for objects of class summary.impactspar.pspatreg
}
\examples{
# See examples for \code{\link{impactspar}} function.
}
\seealso{
\itemize{
  \item \code{\link{impactspar}} Compute direct, indirect and
    total impacts for continous parametric covariates.
  \item \code{\link{summary.impactspar.pspatreg}} Summary method
    for \emph{summary.pspatreg} objects.
}
}
\author{
\tabular{ll}{ 
  Roman Minguez  \tab \email{roman.minguez@uclm.es} \cr
  Roberto Basile \tab \email{roberto.basile@univaq.it} \cr Maria Durban \tab
  \email{mdurban@est-econ.uc3m.es} \cr Gonzalo Espana-Heredia \tab
  \email{gehllanza@gmail.com} \cr 
 }
}
